fix(endpoint): enable Scala 2.13 cross-build
